3D_models_by_hunyuan/requirements.txt
Flask==2.2.5
python-dotenv==0.21.1
requests==2.32.3
tencentcloud-sdk-python-ai3d==3.0.1207
tencentcloud-sdk-python-common==3.0.1207
